working on collapsing the protocol hierarchy
