--- Sources/SwiftVizScale/Documentation.docc/TickScale.md
@@ -1,14 +1,0 @@
-# ``SwiftVizScale/TickScale``
-
-## Topics
-
-### Inspecting Tick Scales 
-
-- ``SwiftVizScale/TickScale/desiredTicks``
-
-### Generating Ticks for a Scale
-
-- ``SwiftVizScale/TickScale/ticks(rangeLower:rangeHigher:)-11uxe``
-- ``SwiftVizScale/TickScale/ticks(rangeLower:rangeHigher:)-224d8``
-- ``SwiftVizScale/TickScale/ticks(rangeLower:rangeHigher:)-68c4d``
-- ``SwiftVizScale/TickScale/tickValues(_:from:to:)``
